move tag_e_bins to r2s
